move httbin_ssl_context fixture into the one place it's used
