+++ tests/integration/test_aiohttp.py
@@ -1,4 +1,6 @@
 import logging
+import os
+import ssl
 import urllib.parse
 
 import pytest
@@ -10,6 +12,18 @@ aiohttp = pytest.importorskip("aiohttp")
 
 
 from .aiohttp_utils import aiohttp_app, aiohttp_request  # noqa: E402
+
+
+@pytest.fixture
+def httpbin_ssl_context():
+    ssl_ca_location = os.environ["REQUESTS_CA_BUNDLE"]
+    ssl_cert_location = os.environ["REQUESTS_CA_BUNDLE"].replace("cacert.pem", "cert.pem")
+    ssl_key_location = os.environ["REQUESTS_CA_BUNDLE"].replace("cacert.pem", "key.pem")
+
+    ssl_context = ssl.create_default_context(cafile=ssl_ca_location)
+    ssl_context.load_cert_chain(ssl_cert_location, ssl_key_location)
+
+    return ssl_context
 
 
 def run_in_loop(fn):
